fix(helpers): Fixing helpers which give unresolved symbols while trying to import.
